"7:50am. I got up 20m and am just chilling. There is not much to do this morning, so I'll start the lectures soon.
8am. https://www.pixiv.net/member_illust.php?mode=medium&illust_id=75297212

This image of Miku doing necromancy is pretty fascinating. It would be interesting if that is where the story goes, but I doubt it. I do doubt is that this will be the last Symphogear season. There is zero reason to end it after S5 and a lot of reason to continue it. S1 had some ambitions of it being a serious story but that was quickly abandoned and since then it has been a XCOM/idol/magical girl hybrid show.

8:05am. Well, let me start. I've been chilling too much.

8:35am. 18m in. Nothing quoteworthy yet, but in yesterday's lecture, Harper saying that I do not need to study algebraic topology to understand this does take a load of my mind. I really like how slowly he is going here and the story telling he does.

Ah, let me quote something.

"In the classical literature this is called modus poenens (function application) because who knows, someone along the way classified these things and used the latin words for them. They must have been British. They wrote these things down.

Typically you do not get this (the introduction rule for functions). This is the innovation of Goentzen. That is important.

Funnily enough as natural as it is and as beautiful as it is in my opinion. All the textbooks they rarely - if they discuss natural deduction it is in chapter 41 and they say 'Oh you know, there is this other thing.' And I think it is kind of unsatisfactory.' But anyway, that is the way it is."

I am actually learning quite a bit from this. Since I do not know any, I had no idea that standard logical practice was so arbitrary and removed from computational concerns."