"9am. 24m in. Ah, so this is weakening. I heard this a few times before, but never had it explained. It is simple enough.
9:10am. Now that I think about it, weakening is not valid is Spiral due to the way join points are handled.

9:55am. ">>14185
> Also good luck explaining to a 9001 IQ AI why it should care about you when you don't care about the cow that got killed to make your hamburger.
As if a being of such intelligence will care about the golden rule. It does not take much to reveal how contradictory it is. Just the idea of an imperfect self improvement loop breaks it.

The thing people need understand that having to explain oneself to God is already a failure mode from which there is no getting out. As soon as you imagine such as scenario you've already lost. There is no such thing as making an exchange with a being of such caliber and human morality is deeply rooted in the experience of being a human.

If God exists, then he already understands this and is one of the reasons why he does not hang out with animals. He knows that any interaction is getting in the way of their evolution. Put it another way, if he does decide to interact, you can be sure that it will be with hostile intentions.

The universal silence is how he shows his kindness."

I wrote this one. I should be watching the lecture, but I am taking a mini-break. Let me get back to it.

10:55am. 1:20h lecture took me 3h somehody. Just why is that?

...Oh, it seems the reply I wrote above did not get posted. It seems solving the capcha is not enough.

Yeah, because of things like that. Forget it.

I'll add a little bit to it later and then try again.

Let me write the bit on control.

"If you think about it, once you did deeper below the surface you will find that the essential basis of morality is the notion that one can control others by controlling oneself. It is a social contract.

The basis of the golden rule is such an social ability that humans possess as social creatures.

I find it absolutely ridiculous that some people assume that they could control a God just because they are nice to cows. That is lazy to the utmost and is deserving of punishment. It is arrogance of the highest degree. Those who want to control Gods should at the very least aim to become one themselves."

11:05am. Yeah, this would be a worthy addition to the reply. I'll post it later.

Right now let me start the next lecture. I want at least 3 per day.

The stuff in this one with all the material on ordering was boring. I remember doing the relations in Coq, but I've forgotten all about that by this point.

I am not sure why he is even teaching this."