"12:45pm. Let me resume. I want to hit my goal of 3 lectures for the day. I can chill after that.
12:50pm. 20m in. Damn this is so boring. I can barely listen to this.

Ah, let me extend the break. I'll continue this later.

1:30pm. Let me resume. I recharged for a bit. Let me clean out the one I am watching and the one after it. Surely I should be able to manage this much. I really should backtrack to the previous lecture though and see why exactly he introduces the orderings.

1:40pm. In 28m. He mentions 5 phases of accepting constructive mathematics by Andrej Bauer. I think I saw a paper by him on that.

http://math.andrej.com/2016/10/10/five-stages-of-accepting-constructive-mathematics/
Five Stages Of Accepting Constructive Mathematics

1:55pm. Yeah, it is just as complicated as I remember it. I can't understand a thing in it. Let me resume the lecture instead.

2:40pm. 1:06h in. "A lot of the time students enjoy programming much more than they enjoy math is because you can run the program after you finish it!" He really exclaims this part.

"With a proof, you can say - here is the proof. Okay, that can be satisfying. But with a proof you can't do anything. But if you have a proper treatment of logic in the style I am doing, you are actually going to have dynamics of proof and you are going to be able to run those proofs."

2:50pm. Done with the lecture. Let me take a look at transitivity for IPL once again. It seems that in this course, I will finally learn some category theory.

2:55pm. Let me take a little break to read Baki and then I will watch the third lecture for the day. By the pace I am going, it does not seem like there will be time for the fourth one.

3:05pm. Let me resume. For the past 4.5 years, the way I've gone through my studies was always like this. I keep hoping that I'd get the strength to enjoy math more, but I always have to force myself to focus. I just can't grow up in this area. I guess this is one of the cost of having a mediocre talent in this. But thankfully math is simpler than programming - if I hadn't the ability for programming then I'd really be hosed. Given how time consuming programming is, boredom would definitely kill me there.

Here I can sort of make progress if I have somebody knowledgeable to fill me in.

But thankfully math comes in limited quantity, and it is really heartening that I do not need to study homotopy theory or category theory as long as I know type theory.

Let me start lecture 4.

3:40pm. 26m in. Oh boy, it is just as I said.

"...I can compute with it like you did in school and deduce that m must be equal to n. When you learned learned polynomials in school you probably found it really fun to like manipulate polynomials when first learned about it. And the way you did it, you wrote down one polynomial and you showed it is equal to another one. Or you took a polynomial and you plugged in a number and you showed that it is equal to 17 or whatever it is. So it is just a computing process.

You wouldn't be sitting in this chair if you did not enjoy that."

Oh boy...I am not exactly having the time of my life here. In fact I am really bored here.

It is just cruel the way mediocrity punishes itself. One day I will erase mediocrity from existence.

3:45pm. There is no point on complaining about it. All I can do is keep going forward on sheer willpower."