"4pm. 37m in. Did he introduce Heyting algebra so he could draw a parallel between it and univalence?
I might need to rewatch those earlier lectures when I am done with this...

Or maybe he did it specifically to demonstrate what the properties of transitivity and symmetry and anti-symmetry imply?

Well, let me resume, it is not that important.

4:45pm. 1:13h in. Only 5m left. I had no idea that `de Bruijin` is spoken as `de Braun`.

4:55pm. Done with that lecture. I really was not paying attention for the last ten minutes. Should I take this as a sign or should I soldier on?

Let me soldier on. I will take a break later. 6pm is my limit regardless of when I wake up.

I'll aim to watch it for half an hour and then call it a day.

5:25pm. https://mathematicswithoutapologies.wordpress.com/2015/05/13/univalent-foundations-no-comment/

I sort of skipped the long exchange in the comments here last and now I am reading it.

Let me continue the lecture for 10 more minutes and then I am done. I really had enough of it. I kind of want to read these comments instead.

5:30pm. No I can't go on anymore. Let me stop at 17m into the fifth lecture. I am done for the day here.

Let me read those comments as the outro for the day."