"11:35am. Let me resume. One more lecture and then I think I will call it a day early here.
12pm. Damn this is tedious, let me take another break.

1:20pm. 56m in. Only 17m left in the lecture. This part where the type of the branch can depend on the value like in dynamic languages is pretty mindblowing. In Spiral that could only be done by specializing at compile time.

1:30pm. 9m left. This is tedious. I am barely following it at this point, but thankfully I am nearly done with it. Yeah, 3 lectures per day are my limit.

At this point I'll have time leftover, but I've been at this since 7am so I have nothing to be ashamed of.

1:45pm. I am done. 3 lectures out of the way. The next one will be lecture 8, so I have 16 lectures left to go.

1:50pm. I need to start winding down. Tomorrow I will take a look the review again.

Maybe I should write more in it, but I do not feel like saying anything else at this juncture even though I have things in mind that would be interesting. Just declaring my determination as usual is good enough.

Hopefully by the end of the year the drudge work of doing simple exercises will be done with and I will have a decent and up to date understanding of the latest in type theory. That will allow me to start gaining real insight in the field of mathematics.

In the 2015 [math thread](https://mathematicswithoutapologies.wordpress.com/2015/05/13/univalent-foundations-no-comment), those field medalists downplayed technical skill, but I know that Terry Tao was a participant in the math olympiad before becoming what he is today. I am really so far from what he was even in high school, and working on my technical skills in particular will be a major theme for me in the next year or two.

Similarly how I did with staged functional programming, I need to completely internalize the art of programming with dependent types. Once I do that, it will allow me a much deeper level of insight for the creative work that I need to do.

It would have been great if instead of Logo and Basic, I was studying dependently typed programming back in elementary school. That would have been a far better use of my time back then, but I guess even at my age it is never too late to get familiar with the fundamentals. If even the greats still have not mastered the foundations of mathematics then if I want to get good at math, I should pay my respects to it by practicing the fundamentals.

Fundamentals are something that touches all the other areas and will act as a bonus to my ability in everything I try. I truly believe in doing the basics first.

2:20pm. I feel like chilling for a while now. I probably won't make any commits for the remainder of the day. Either way, I'll resume this tomorrow."