"2:05pm. I woke up 2h ago or so and have just now finished breakfast. I am finally ready to start. I really should have finished the thing earlier than 7am, but that is how it turned out.
I was worried whether I'd fall asleep at all once I get up because I have had issues with that sometimes unless I keep a regular schedule, but it turns out not sleeping for 24h is all it suffices to induce rest. I'll be going tomorrow and that should be it till new year.

2:15pm. Well, that is about that. Now I am just surfing social sites and getting brainwashed again.

Political thought used to be so fun, but now it is just so dull. I need to replace it, but what other sort of entertainment options are open to me? As I grow older, maybe I be inevitably to be drawn to math simply because it is a different kind of game even if I did not have a explicit life purpose for pursuing it.

2:20pm. Let me start lecture 8. I'll aim for one lecture today. I'll post the review either after that or tomorrow morning depending on how I feel after that. I do not want to post it until I get over my grogginess, but right now I am just too hung over to think properly."