"2:55pm. What is the point of lecture 8? 9m in and so far all he has been doing is retread old territory, the teach assistant I mean.
Today is a mess. Well, every 23rd June is a mess in a way, but this one is exceptional because this year the party both started a few hours later and laster a few hours longer as usual so by the time I recover it will be 8-9pm. I'll do the blog post tomorrow. Today let me just focus on recovery and seeing if I can get two lectures out of the day.

Let me start skimming this one.

3:15pm. Focus me, let me just go through the lecture instead of wasting time.

3:50pm. 30m in. Harper comes in after the 11m mark.

The stuff here on Id elimination is interesting. In one of the past lectures Harper did say that the Id type is the most mysterious part of type theory and that even he himself still does not have a full grasp on it. These rules definitely are illuminating.

3:55pm. 34m in. I can't really follow the inference rules directly though.

```
J : (X : 𝓤 ̇ ) (A : (x y : X) → x ≡ y → 𝓥 ̇ )
  → ((x : X) → A x x (refl x))
  → (x y : X) (p : x ≡ y) → A x y p
J X A f x x (refl x) = f x
```

The Agda version of J is really much more straightforward than the thing he is writing on the blackboard.

4:35pm. 1h in. Quote time.

"(on later proofs depending on the details of the previous ones) It's a bit - from a programming point of view that kind of thing is not very robust. Because you are introducing an inter-module dependencies. The proof of this theorem depends on the exact proof and not the fact that the theorem is true.

That's kind of a crappy thing to be honest. And the modularity issues about these kind of formalized developments in type theory are not fully methodologically worked out. People really depend at this stage on 'I have a linear sequence in my library and every damn thing I said before to a first approximation, every last thing depends critically - the next thing depends critically on all of it.' That typically happens. Which is not like a piece of code that you usually write. You usually strive to ensure that's a library and I don't give a damn how it's implemented and in fact I do not even have it around when I am writing my code and compiling my code. I just link it later.

That kind of thing is not well developed in the mechanized, dependent reasoning world. And one manifestation of that is that the definitional equivalences that hold of a proof can affect the well formedness of a subsequent proof."

4:50pm. 1:07h in. It is really a struggle to keep my eyelids up at this point.

I do not think I am going to be able to bear another lecture. Right now he is trying to define `ap` using complicated substitution rules.

```
ap : {X : 𝓤 ̇ } {Y : 𝓥 ̇ } (f : X → Y) {x x' : X} → x ≡ x' → f x ≡ f x'
ap f {x} {x'} p = transport (λ - → f x ≡ f -) p (refl (f x))
```

The Agda implementation on the other hand is quite clear. This is from the Intro to HoTT/UF book.

5pm. 1:13h in. Here he explains the fibers.

5:15pm. Done with the lecture.

I almost feel like starting the next one here. It was really boring at one point, but then I got some interest towards the end.

5:20pm. I guess I will give it a try.

6:25pm. 41m in. I am going to stop here in lecture 9. I guess doing 1.5 lectures for today should be fine. It will make it easier for me tomorrow.

6:40pm. https://mangadex.org/title/32154/girls-frontline

I wouldn't put this as one of the top titles, but I like the atmosphere of the thing. It is quite interesting so it has me following it. The action is mediocre, but if the anime turns out to be nitty gritty and with decent budget it could have potential. The PV does do a dumb fakeout into a comedy sketch, but maybe the direction will change into something that has potential to be big rather than just pandering to the mobile crowd.

6:45pm. Tomorrow while I will go out during the day again, I doubt it will be a 9pm-7am thing like yesterday. Most likely it will be 4-9 or 5-10 thing, meaning once that this is done I will able to get back into the swing of things after this mini-vacation. Hopefully my sleeping schedule was not entirely disrupted because of this.

I'll plug the review and hopefully have time to fulfill my lecture quota for the day.

Going through these lectures takes me forever and I am not getting that much out from them, so I will be glad to be done with this drudgery by the end of the month. I want to get back into programming.

I have a bit more than just the lectures from 2013 that I am going now though. I also want to go through the 5 lecture series on computational type theory by Harper from OPLSS 2018. I lost my bearings on the last two eps and I want to test myself after all this is done to see if I can digest them.

6:55pm. I think that after I am finished with all the SF volumes that I will have some real skill in dependently typed programming at my disposal which will allow me to properly dive into the Coq and Agda libraries and study them. Slowly the curtain will rise on the real deal. From the programming language perspective, the passage of time is on my side as the development on cubical type theory will directly feed into my own capabilities.

Even it bears keeping in mind even though I am not expecting the ML field to do much on its own, there is still a chance that it might defy expectations and make a significant breakthrough while I am doing my own stuff.

I will leave nothing to chance. The ML field will keep getting better, I will keep getting better and the PL field will keep getting better. I have every intention of making all the advantages my own. Doing it like this is far superior than just waiting for the thing to drop into my lap. Assaulting the issue from numerous different angles is how the winners do it."