"9:50am. 1:02h. I actually really like how opinionated he is. Here he compared proof relevance and irrelevance, compared them to static and dynamic typing and declared static typing to be strictly superior.
"If you say that on Reddit you will get flamed to death and nonetheless it is true."

I like the story times. It really does make the material come to life. That Agda book in comparison is really missing out on the narrative aspect.

9:55am. http://www.cs.cmu.edu/~rwh/courses/chtt/

This seems to be a more recent course by him. I am not sure if it has yet started to run.

10:05am. https://www.youtube.com/watch?v=0fR6FVQ0SP4

Here is the RedPRL demo.

http://www.cs.cmu.edu/~rwh/talks/POPL18-Tutorial.pdf

Right now I am reading this, but I really should be watching the lectures instead. Let me get back to it. Just close this down.

10:15am. 1:15h in. "Only identifications are the refls."

Well yes, this is the main point as to how I've been thinking about type theory. I guess not having it be so would create a lot of problems. I really am interested in how it will turn out to be in HoTT.

10:20am. 1:18h in. "And life is very good. If you want to formalize any kind of conventional mathematics you run into every single day this (ETT) is the way to go."

He also says that NuPRL is based on ETT and Coq is based on ITT.

"The experience until like couple of years ago (from 2013), the experience of doing things this way was every single person ends up complaining bitterly about the obnoxious need to pass around all these proofs of identities that are only reflexivities and it gets really, really tedious. Whereas NuPRL never had that problem."

10:35am. Let me start the next lecture. I managed to get through the first half.

10:50am. 18m in. This is interesting. He says that ITT introduced all the machinery needed for the types be infinity groupoids, but could never realize it as it had no mechanism for equality other than reflexivity.

11am. 25m in. "HoTT = ITT + axioms that introduce new identifications amongst elements of a type"

11:05pm. 27m in. Here is really explaining everything properly and I am understanding it. Great. Simply wonderful.

11:10am. 29m in. My intuition is really telling me strongly that I am going to need this once I start messing with probability distributions.

12:35pm. Done with both lecture 10 and breakfast. The next one is lecture 11.

12:50pm. Focus me, let me just do this."