"1:05pm. Ah, let me take a break. I keep forgetting that it is Monday. Let me see if Tog and Satanophany are out.
I actually really do like the whole feel of Tog. Even as bland as Baam used to be, he is pretty cool once he starts fighting. The young Jahad fight was kino. The writer is actually pretty red piled all together.

You can see it in Kalavan's backstory - he is a descendant of a community which had a tradition of marrying the women to death match winners, which as a result makes him very strong by blood. A more sensitive author would not consider showcasing eugenics in such a positive context.

A lot of people do not really understand Rachel. I admit I myself was confused at the end of the first arc as to why she pushed Baam away, but in the interim given how stacked fate is in Baam's favor, had I been in her shoes, I'd have pushed the shitter off myself. She knows the world is full of shit and is annoyed at herself for not being able to take advantage of that insight. Baam is pretty unlucky though to have picked such a strong willed waifu. But immortality is a thing in that universe so it will be interesting to see how their relationship evolves as the time goes by.

I really overdid it with the Tog references in the last update. I have no idea what I was thinking when I started talking about thorns. I suppose anything could be justified with the words 'pleb filtering', though I do not think anybody is really reading those updates.

So I have to be doing it for myself. It would really be cool if the story is leading into Baam purging the whole tower as a part of manifesting his full powers. It would take the author a lot of guts for it to be that way.

Hmmm...now that I think about it, since I am talking about SIU being red piled, if I was going to point out an example of a very blue piled author, that would be Ken Akamtsu.

Somehow that is how the whole backstory by Eva comes out in my mind with the whole 'violence breeds violence' and 'chains of hatred' cliche he continually pushed forward. And then in the latest chapters when he had Touta explain that immortality for everyone would take too much energy so much that even tapping the energy directly from sun's core would not be enough that clinched it in my mind.

Akamantsu might be a master of fanservice, but a new culture is needed. The one that currently exists is just too tepid for my taste.

Let me have a proper break. Somehow I spent too much time just thinking about this."