"1:50pm. Well, let me just close the thought in the last entry that really most of the political questions so to speak are really salad dressing. The only really important aspect of political affiliation is the genocide propensity.
Left wing or right wing makes little difference. But also in addition to that it must be said that the presence of computers makes a decisive difference as they are the ones which will allow for tiered reality and elimination of costs both social and material that would be otherwise be associated with genocide.

I mean, real violence is bad, but virtual violence is not. What exactly happens to such a moral judgment in an universe where everything is virtual? Classical thoughts of both wings will get clipped by this question. For that reason I do not really want to be right, left or center. I want to affiliate myself as much as possible with the future.

Let me resume here.

Maybe 3 lectures will be too much, but I should be able to manage 2.5.

2:35pm. 26m in. Roll-over-path. This is a new concept. I do not think I saw this in the Agda book. This is one of the things I've been wondering about.

3pm. 48m in. He is being pretty convincing at telling me why not to study set theory.

3:20pm. 1:07h in. I thought I understood equivalence, but it seems I do not. I really thought it would be `f (g x) = x` and so on, but it turns out that `=` is equivalence...

I never saw that in the Agda book.

This is quite strange.

3:30pm. 1:13:50h in. I am really confused as to why he is calling this equivalence.

3:40pm. Done with the lecture. Let me take a break here. If I have time in the day once I get back, I'll try to get through another lecture.

To answer my earlier question, the reason why he started calling it equivalence is because that statement is valid under the equivalence axiom. He just got a bit too hasty with it.

4:20pm. I'll resume in a bit...no, let me do it now. As much as I want to I should leave manga for later. I still have steam for one more lecture, and relaxing that would make me lose the chance.

Let me start lecture 12.

4:40pm. 8m in. Funny how often inverses come up. In Bayesian ML the inverse is the posterior distribution and it came up when doing credit assignment in Q Zap and LSTD. It comes up in KFAC and second order methods.

There is a sense that reversing computation is of definite importance for learning, so it is interesting that it a vital part of equivalence.

4:45pm. I am interested in why the quasi-inverse is logically equivalent to equivalence.

Let me stop the lecture for a bit so I can look it up in the Agda book."