"4:50pm. No I can't find it easily in the Agda book. And the HoTT book is complicated. Forget it, let me just go through the lecture. The specific proof can wait.
6pm. 1:18h. Only two minutes left. There is a funny bit towards the end.

"So by the way, this dependent elim. This has been around since 1971 in Martin Lof's original type theory. This is what in broken form is called a GADT in Haskell. In Haskell they have this horrifically hacky, kludgy, broken thing involving polymorphism that is attempting to express the idea that when you do a case analysis when you are in a branch you flow the information about which branch you are in into that branch. Ignore GADTs. It's like...garbage. The right thing is like dependent types. Just so you know."

6:05pm. Done with lecture 12. I am a damn hero to get through 3.5.

I'll call it a day here. I really understand pretty much everything in the past 12 lectures. I think I'll be able to learn something of real substance from this course, though I won't get the chance to use it until I get through all the homework first which will take me a few months. Given how long it took me for the first SF volume - 1.5 months, I do not think I'll be able to go through the other 3 + the Agda one any faster."