"11:40am. 56m in. Let me just say a few things about the identity type.
What he is writing on the board is hard to follow due to boredom, but I get the gist of it. He is struggling a lot to explain HoTT without necessarily having a computational interpretation.

But this claim that the identity type is special...

There is a bunch of stuff that Spiral does that ITT necessarily does not, that might be possible to use to extend it. One thing is the notion that the computation done at the type level need not necessarily be computable.

Furthermore, I've always found it surprising the the identity's constructor `refl` actually carries with it an item rather than being a blank. Suppose you have something like `Id 3 3`. You do not need an item here as it can be manifested from information directly on the type level. Even for thing like `Id n n` where the value is not obvious from type, that erasure behavior could be recovered by nesting JIT compilation.

Now I do not know whether these concepts could be extended so it covers the univalence axiom and higher inductive types, but it might be worth a try. Still what I am proposing here is at least an order magnitude more difficult than the creation of Spiral, so I doubt I will be doing it. Maybe it will be enough for me to merely master ETT?

11:50am. Let me finish the lecture, only 19m more left."