"7:20am. I got up a bit ago.
7:35am. I am still quite groggy, but how about I start the lecture?

7:50am. Let me do it.

8:10am. http://www.cs.cmu.edu/~rwh/courses/chtt/pdfs/week1.pdf

I'll go through the notes for the CTT course later.

8:40am. 27m in. This is really tedious. Really, the course on higher TT really should have a computational interpretation. I learned a lot from the course and will finish it, but as it is right now the thing is a joke.

I sort of get into the mood to study every once in a while, but once I get back into the swing of things, I won't stop until I've dealt with all the SF volumes.

9:10am. 57m in. Ok, this definitely caught my attention a little. I can't even imagine how `f : A -> B` implies that `qinv f ~= (fun x => x = x)`. This is insane.

9:25am. 1:10h in. This is just a random tidbit here, but there might be something profound in the equivalence between the sum of a fiber and the type itself.

9:40am. Finally done with the lecture. It was pretty boring.

Let me move to the next one.

9:50am. Let me. Focus, just focus.

10:15am. 15m in. I am really not focusing, but by my desire, I am going to get through this.

10:50am. 22m in. Done with breakfast. Let me finish lecture 19 and then I will take a break. Nanoha Detonation subs are out and I just finished downloading it so I am looking forward to watching this.

When I am studying and programming things actively I prefer anime to games as it is harder to get hooked on it to the point I feel like not doing the things that need to be done. Generally this has been true for the past 4.5 years, but it definitely was not true all the way throughout my life.

11:30am. 53m in. This part is absolutely droning. I can't wait until the series is over. 30 more minutes left in this lecture.

12:15pm. The lecture is finally over. This is such a convoluted way to introduce inductive types. I can already feel myself not giving a crap, but I really need to get through this. Just one more lecture today and 3 tomorrow and then I am done with this.

The next level after this should be more straightforward as computational type theory will have a computational basis. And things that compute are easier to understand than things that just describe.

Well, let me take a real break here.

I'll continue in an hour or two."