"12:25pm. https://www.cs.cmu.edu/~rwh/papers/cartesian/paper.pdf "Our type theory is the first two-level type theory with canonicity, and the second univalent type theory with canonicity, after the cubical type theory of Cohen et al. [17]. Like Cohen et al. [17], our type theory is inspired by a model of homotopy type theory in cubical sets [12], and represents n-dimensional cubes as terms parametrized by n variables ranging over a formal interval. However, the fibrant fragment of our type theory differs from Cohen et al. [17] by endowing the interval with less (namely, Cartesian) structure, and defining fibrancy with a substantially different uniform Kan condition. Thus we affirmatively resolve the open question of whether Cartesian interval structure constructively models univalence [18, 22]."
Is this all that is needed for HoTT to be computable.? I guess I'll know after I read this paper and go through the OPLSS lectures.

1:10pm. Let me stop at the 30m mark.

1:20pm. Let me resume the lecture. I want to get this out of the way.

https://news.ycombinator.com/item?id=16088622

Let me just make a comment on mbid's comments.

The reason why I need to do this is because I am seriously suspecting that credit assignment flows from the structure of the forward model. I need to find the right class of models and reverse them - and in addition to this there should be something about the nature of Bayes rule that will make things easier to do this. At first glance it seems that just having it will derive no real benefit in the search for the algorithm, but there should be something there. There should also be something about how to connect disparate models and localize them.

I need something absolutely strict and formal as a language to have a chance of stumbling upon it.

If I were doing it in Spiral, I would never even get close to the answer of that I am sure. When I start implementing regular math and learning algorithms from the ground up, no doubt I will run into difficulties that I wouldn't have in say Spiral. Those stumbling blocks and that failure are actually the keys to solving this.

It really bothers me how the Bayes rule to start with actually requires reals. I do not think credit assignment should be that complicated - it should be more fundamental. Learning is intrinsic to life. It should flow from the notion of computation.

"Hi Ingo, I'm pleasantly surprised to see an actual researcher here."

Where do mathematicians get their haughtiness from, I wonder sometimes.

I've been thinking about the difference between programming and mathematics, but the real difference is between classical and constructive mathematics. This is a distinction that really makes sense. Harper never said this, but the difference between the two is essentially the one between talking and doing.

"This is somewhat true, but I think reasearch in type theory is connected to that in other branches of mathematics (if you want to call type theory maths) very weakly in comparison. There are some connections to category theory, but they are far to weak given that type theory should essentially be a branch of categorical logic -- the one that deals with the initial so-and-so category and its computational properties."

Math that lacks a computational context is nothing more than a figment of imagination.

This might be the first time I've put this notion into words, but as a programmer this is something that I've long since internalized. I've never trusted math one bit, but computational type theories are eminently trustworthy.

Let me resume the lecture."