"https://math.stackexchange.com/questions/1158383/examples-of-the-benefits-of-homotopy-type-theory-for-computer-aided-proofs
Later, later...

1:55pm. 9m in.

2:15pm. 25m in. I am not following this at all at this point. I am just gritting my teeth through it. Either way, I do not think it is important for me to understand this. Let me just go through this so I can earn my badge of accomplishment.

3:20pm. 1:08h in. This is just miserable. Inductive principles were complicated in Coq and they are 10x more complicated when written out like this on the blackboard. Right now I am back in high school sitting at my desk and not paying attention. Just why am I bothering doing it like this? It is not like anyone is forcing me.

Well, I am nearly done with this lecture. One more day of this and I will be done for real with the course. Unlike high school at least this thing is over in 8 days rather than lasting 4 years.

Not all lectures were as boring as the last two though. I am really glad I took the time to go through this the lectures as it gave me a lot of understanding about what type theory is and what the relationship between programming and math is. Without going through this not only would I not know that, but I would be clueless as to the difference between ITT, ETT and other type theories.

Since it lacks an identity type, I can be sure that Spiral's type system is not a type theory, but a more vanilla type system.

3:30pm. 1:15h. I completely wasted my time watching this lecture. I do not even have any idea what those `W`s are supposed to be.

...Just another kind of function application it seems.

3:50pm. Well, I think I will call it a day here. I want to resume where I left of during the break.

I'll read the CTT paper for a bit and check out that question, but that is about that.

Just a bit more and I will be done with these lectures. Just tomorrow. Then the more fun stuff will start."