"7am. I honestly thought that today I would be getting up later than usual so I was quite surprised when I looked at the clock 15m ago. I'll start soon. After the 3 lectures I'll round the day off with the 2 talks - or at least one.
I am really sick of the lectures at this point, but I want to watch something on CTT even after I am done.

7:05am. Let me start this thing.

7:25am. Connection is breaking at random times. I hate the slow mode.

7:35am. 18m in. Finally the subject turns to higher inductive types. I still have no idea how they work and probably won't until I start playing with cubical types, but this should serve as an intro.

8:55am. Done with the first lecture. Let me take a little break and then I'll start the penultimate one.

9:10am. Let me do the second one. I like this pace of two lectures in the morning.

9:30am. 12m in. Let me say a few words.

I am really not obsessed about understanding this at this point. I am fine with math geniuses figuring this stuff out first and then teaching it to me. I know that to get a handle on this it is not enough to just talk about it or listen to others talk, but to actually program with it.

11:25am. Done with both the second lecture and breakfast. Usually I'd be taking a break at around this point, but I just want to get the course done with. I am motivated to put that final video behind me and finally move on to the next thing.

Let me do it. Then I'll reward myself with Choubu no Shinobi raws. I really hope that the thread for the next chapter Himenospia (when it comes out) does not die in 5m.

Actually, the way the last chapter went is actually quite good. Mizu puking at Himeno trying to sacrifice herself is actually completely in her character and yet I completely did not see it coming despite being aware of the perspective that holiness is fake. This is actually surprisingly good writing. Mizu acting as the moral backbone for Himeno right now. Not the kind that says what is 'good', but the one that say what is necessary.

https://boards.4channel.org/a/thread/189865994/tokushusen-asuka-45

Oh, and more of this is out. Sweet.

Well, let me start the last lecture."